requirements.txt
fastapi
uvicorn
pydantic
httpx
requests
streamlit
openai
python-dotenv
tqdm
chromadb
networkx
numpy
torch
transformers
sentencepiece
scipy
chromadb
streamlit
sentence-transformers